Add files via upload
The project is an Airport Management System GUI application using Python Tkinter. It enables users to input and manage passenger and flight information, storing data in a MySQL database. The program includes features such as adding, deleting, and displaying passenger and flight records. It also has a dedicated section for employee information with its own window. The project involves various interconnected tables in the database. The program uses the MySQL connector to interact with a MySQL database and handles exceptions for error handling. Overall, it's a comprehensive application for managing airport-related information with a graphical user interface. The deletion part has some issues,it will be resolved soon.